fixed mix-up between number of rows and the number of columns, changed terminology
The example used 'extent_rows' as the length of a matrix row, which
equals the number of columns. Same for 'extent_cols'. This was confusing
and actually was mixed up in this small example. Now it is changed to
'rows' for the number of rows and 'cols' for the number of columns. This
should be much easier to understand when starting with such simple DASH
examples.