The inline function (dart_base_logging_basename) resides in the header 'logging.h'.
If the code can't be inlined for whatever reason (e.g. on Hazel Hen Machine), the final executable has to have one version of the function to which it may link. This we would have to "instantiate" in just one compilation unit (logging.c file).

extern inline char * dart_base_logging_basename(char *path); // Forces the symbol to be included in the unit.