Move content of `.gitignore` to `.gitmodules`.
Git does not read any full fledged config files from the repository, in
particular any `.gitconfig`. Thus this file is useless, without a
`include.path` statement in any config file Git is reading (i.e.,
`.git/config`). Though it reads `.gitmodules` and as the content of
`.gitconfig` is sub-modules related anyway, it has an effect, if it is in
this file. Do so.